finish feature/sign-up-api
