+++ lib/core/routes/app_routes.dart
@@ -3,13 +3,13 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:quick_mart/core/network/dio_handler.dart';
 import 'package:quick_mart/core/routes/routes.dart';
 import 'package:quick_mart/features/auth/forget_password/views/screens/success_screen.dart';
-import 'package:quick_mart/features/auth/login/data/repo/sign_up_repo.dart';
-import 'package:quick_mart/features/auth/login/views/screens/log_in_screen.dart';
+import 'package:quick_mart/features/auth/sign_in/sign_up/data/repo/sign_up_repo.dart';
+import 'package:quick_mart/features/auth/sign_in/sign_up/views/screens/log_in_screen.dart';
 import 'package:quick_mart/features/on_boarding/view/screen/on_boarding_screen.dart';
 
-import '../../features/auth/login/data/repo/login_repo.dart';
-import '../../features/auth/login/logic/cubit/Auth_cubit.dart';
-import '../../features/auth/login/views/screens/sign_up_screen.dart';
+import '../../features/auth/sign_in/sign_up/data/repo/login_repo.dart';
+import '../../features/auth/sign_in/sign_up/logic/cubit/Auth_cubit.dart';
+import '../../features/auth/sign_in/sign_up/views/screens/sign_up_screen.dart';
 import '../../features/home/home_screen.dart';
 import '../../features/splash/screens/splash_screen.dart';
