+++ lib/features/auth/sign_in/sign_up/data/model/sign_up_model.dart
