+++ lib/features/auth/sign_in/sign_up/logic/cubit/Auth_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
-import 'package:quick_mart/features/auth/login/data/model/sign_up_model.dart';
+import 'package:quick_mart/features/auth/sign_in/sign_up/data/model/sign_up_model.dart';
 
 import '../../data/model/login_model.dart';
 import '../../data/repo/login_repo.dart';
