+++ lib/features/auth/sign_in/sign_up/views/widgets/term_text.dart
